+++ src/drivers/OPH_APPLY_operator.c
@@ -1612,7 +1612,7 @@ int task_execute(oph_operator_struct * handle)
 
 	int l;
 
-	int num_threads = (oper_handle->nthread <= oper_handle->fragment_number ? oper_handle->nthread : oper_handle->fragment_number);
+	int num_threads = (oper_handle->nthread <= (unsigned int) oper_handle->fragment_number ? oper_handle->nthread : (unsigned int) oper_handle->fragment_number);
 	int res[num_threads];
 
 	//In multi-thread code mysql_library_init must be called before starting the threads
@@ -2301,8 +2301,8 @@ int task_destroy(oph_operator_struct * handle)
 		//Delete fragments
 		int num_threads =
 		    (((OPH_APPLY_operator_handle *) handle->operator_handle)->nthread <=
-		     ((OPH_APPLY_operator_handle *) handle->operator_handle)->fragment_number ? ((OPH_APPLY_operator_handle *) handle->
-												 operator_handle)->nthread : ((OPH_APPLY_operator_handle *) handle->operator_handle)->fragment_number);
+		     (unsigned int) ((OPH_APPLY_operator_handle *) handle->operator_handle)->fragment_number ? ((OPH_APPLY_operator_handle *) handle->
+												 operator_handle)->nthread : (unsigned int) ((OPH_APPLY_operator_handle *) handle->operator_handle)->fragment_number);
 
 		if (((OPH_APPLY_operator_handle *) handle->operator_handle)->fragment_id_start_position >= 0 || handle->proc_rank == 0) {
 			if ((oph_dproc_delete_data(id_datacube, ((OPH_APPLY_operator_handle *) handle->operator_handle)->id_input_container,
