+++ src/drivers/OPH_REDUCE_operator.c
@@ -1000,7 +1000,7 @@ int task_execute(oph_operator_struct * handle)
 
 	int l;
 
-	int num_threads = (oper_handle->nthread <= oper_handle->fragment_number ? oper_handle->nthread : oper_handle->fragment_number);
+	int num_threads = (oper_handle->nthread <= (unsigned int) oper_handle->fragment_number ? oper_handle->nthread : (unsigned int) oper_handle->fragment_number);
 	int res[num_threads];
 
 	char _ms[OPH_COMMON_MAX_DOUBLE_LENGHT];
@@ -1330,8 +1330,8 @@ int task_destroy(oph_operator_struct * handle)
 		//Delete fragments
 		int num_threads =
 		    (((OPH_REDUCE_operator_handle *) handle->operator_handle)->nthread <=
-		     ((OPH_REDUCE_operator_handle *) handle->operator_handle)->fragment_number ? ((OPH_REDUCE_operator_handle *) handle->
-												  operator_handle)->nthread : ((OPH_REDUCE_operator_handle *) handle->operator_handle)->
+		     (unsigned int) ((OPH_REDUCE_operator_handle *) handle->operator_handle)->fragment_number ? ((OPH_REDUCE_operator_handle *) handle->
+												  operator_handle)->nthread : (unsigned int) ((OPH_REDUCE_operator_handle *) handle->operator_handle)->
 		     fragment_number);
 
 		if (((OPH_REDUCE_operator_handle *) handle->operator_handle)->fragment_id_start_position >= 0 || handle->proc_rank == 0) {
